+++ source/renderer/app/features/discreet-mode/context.tsx
@@ -1,4 +1,4 @@
-import React, { useState } from 'react';
+import React, { useEffect, useState } from 'react';
 // @ts-ignore ts-migrate(2305) FIXME: Module '"react"' has no exported member 'Node'.
 import type { Node } from 'react';
 import { merge } from 'lodash/fp';
@@ -9,6 +9,7 @@ import {
 import { useLocalStorageFeature } from '../local-storage';
 import { DiscreetMode } from './feature';
 import { DiscreetModeApi } from './api';
+import { useAnalytics } from '../../components/analytics';
 
 export const discreetModeContext = React.createContext<DiscreetMode | null>(
   null
@@ -16,19 +17,28 @@ export const discreetModeContext = React.createContext<DiscreetMode | null>(
 interface Props {
   children: Node;
 }
+
 export function DiscreetModeFeatureProvider({ children }: Props) {
   const localStorageFeature = useLocalStorageFeature();
+  const analyticsTracker = useAnalytics();
+
   const [discreetModeFeature] = useState<DiscreetMode>(() => {
-    const feature = new DiscreetMode(new DiscreetModeApi(localStorageFeature));
-    // @ts-ignore ts-migrate(2339) FIXME: Property 'daedalus' does not exist on type 'Window... Remove this comment to see the full error message
+    const feature = new DiscreetMode(
+      new DiscreetModeApi(localStorageFeature),
+      analyticsTracker
+    );
+
     window.daedalus = merge(window.daedalus, {
       features: {
         discreetModeFeature: feature,
       },
     });
+
     return feature;
   });
+
   useFeature(discreetModeFeature);
+
   return (
     <discreetModeContext.Provider value={discreetModeFeature}>
       {children}